VPN-5049 - Move sharePlainText back to VPNUtils (#7460)
* Move sharePlainText back to VPNUtils

* Update android/vpnClient/src/main/java/org/mozilla/firefox/vpn/qt/VPNUtils.kt
